CRUD subject & grade

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -20,7 +20,7 @@ class CreateUsersTable extends Migration
             $table->timestamp('email_verified_at')->nullable();
             $table->string('password');
             $table->enum('role', ['admin', 'pj', 'asisten']);
-            $table->integer('assistant_id');
+            $table->bigInteger('assistant_id');
             $table->string('photo');
             $table->date('join_date');
             $table->rememberToken();
